Added Libnetcdf and netcdf4 to the repo
Omnia packages which depend on it will downgrade/fall behind the longer we are not on conda-forge.
I have added these 2 packages to help keep some of our packages up to date.

Additionally added some `extra` information on other packages I personally have added to make it clear who the point-of-contact should be (me) for the builds as they exist on omnia.