-y option no longer valid in homebrew-cask?
Not sure what is wrong here, but something changed either in Homebrew or Homebrew-cask that made it so the -y option isnot accepted in `brew cask` anymore. I don't know if this is a bug or intentional yet and no one has raised an issue on either github, so I am writting in a || to try and error trap a bit for both cases.