Unpin OpenMM's Sphinx dep.
I think we can safely unpin OpenMM's sphinx dependency from 1.5.6. I have tested locally and it appears to work.

We also need to unpin sphinx if we want to build on Python 3.7 since they are not making the old versions of sphinx on Py37.

We should not need to trigger new builds for Py 27, 35, or 36 since no code changes here, just the way the docs are compiled.